docs: migrated `provider` docs snippets (#3357)
* docs: moved provider options docs snippets

* docs: moved provider instantiation docs snippets

* docs: moved provider getBaseAssetId docs snippets

* docs: removed un-used provider.test.ts docs snippets

* docs: moved provider getCoins docs snippets

* docs: removed /guide from snippet imports

* docs: moved provider getResourcesToSpend docs snippets

* docs: moved provider getBalances docs snippets

* docs: moved provider getBlocks docs snippets

* docs: moved provider getMessages docs snippets

* docs: moved provider getMessageByNonce docs snippets

* docs: moved provider getMessageProof docs snippets

* docs: moved provider getTransactions docs snippets

* feat: export `PageInfo` from `Provider` class

* docs: moved provider pagination docs snippets

* chore: changeset

* docs: finalize the migration of provider snippets

* chore: fixing tests

* docs: fixed getMessageProof tests

* chore: lint
