+++ .changeset/hip-fans-scream.md
@@ -1,0 +1,5 @@
+---
+"@fuel-ts/account": patch
+---
+
+docs: migrated `provider` docs snippets
