+++ apps/docs-snippets2/src/provider/provider-operations.ts
@@ -1,0 +1,15 @@
+// #region operations
+import { Provider } from 'fuels';
+
+import { LOCAL_NETWORK_URL } from '../env';
+
+// Create the provider
+const provider = await Provider.create(LOCAL_NETWORK_URL);
+
+const chain = await provider.operations.getChain();
+const nodeInfo = await provider.operations.getNodeInfo();
+// #endregion operations
+
+console.log('provider', provider);
+console.log('chain', chain);
+console.log('nodeInfo', nodeInfo);
