+++ apps/docs/src/guide/provider/index.md
@@ -4,6 +4,6 @@ The [`Provider`](../../api/Account/Provider.md) lets you connect to a Fuel node 
 
 All higher-level abstractions (e.g. [`Wallet`](../wallets/index.md), [`Contract`](../contracts/index.md)) that interact with the blockchain go through the `Provider`, so it's used for various actions like getting a wallet's balance, deploying contracts, querying their state, etc.
 
-<<< @/../../docs-snippets/src/guide/provider/provider.test.ts#provider-definition{ts:line-numbers}
+<<< @/../../docs-snippets2/src/provider/provider-instantiation.ts#provider-instantiation{ts:line-numbers}
 
 You can find more examples of `Provider` usage [here](./querying-the-chain.md).
